feat(desktop): integrate image support and enable asset protocol
The asset protocol is necessary to let images load from frontend
via links like asset://localhost//home/sam/hey.png. I added $HOME and /usr
as I consider that this is the only place where we would like to allow loading
images. This could easily be changed later, especially if this is not working on Windows.